+++ frontend/package.json
@@ -56,7 +56,7 @@
     "tailwindcss": "^3.4.1",
     "typescript": "^5"
   },
-  "packageManager": "yarn@4.6.0",
+  "packageManager": "yarn@4.9.1",
   "engines": {
     "node": ">=20"
   }
